Create Postorder Interative (S2).c
 (postOrderIterativeS2) Write an iterative C function postOrderIterativeS2() that
prints the post-order traversal of a binary search tree using two stacks. Note that you should only
use push() or pop() operations when you add or remove integers from the stacks. Remember
to empty the stacks at the beginning, if the stacks are not empty.
The function prototype is given as follows:
void postOrderIterativeS2(BSTNode *root);
Following is the detailed algorithm:
1) Push root to first stack.
2) Loop while first stack is not empty
2.1) Pop a node from first stack and push it to second
stack
2.2) Push left and right children of the popped node to
first stack
3) Print contents of second stack